added BaseChart component w/ default props
