+++ src/interface/others/FooterChart.tsx
@@ -1,25 +1,12 @@
 import React from 'react';
 
 import { TopLevelSpec, Config } from 'vega-lite';
-import { VegaLite, VisualizationSpec } from 'react-vega';
+import { VisualizationSpec } from 'react-vega';
 import { AutoSizer } from 'react-virtualized';
 
-export const defaultConfig = {
-  autosize: {
-    type: 'fit' as const,
-    contains: 'padding' as const,
-  },
-  background: '#0000',
-  padding: 0,
-  view: {
-    stroke: null,
-  },
-  axis: {
-    domainColor: 'gray',
-    labelColor: '#e9e8e7',
-    tickColor: 'gray',
-  },
-};
+import BaseChart, { defaultConfig, formatTime } from './BaseChart';
+
+export { formatTime };
 
 export type Spec = Omit<TopLevelSpec, "data">;
 
@@ -28,10 +15,6 @@ export interface Props {
   data: any;
   config?: Config;
   height?: number;
-}
-
-export function formatTime(field: string = 'datum.timestamp') {
-  return `if(${field} / 1000 >= 60, toString(floor(${field} / 60000)) + 'm ', '') + toString(floor(${field} / 1000) % 60) + 's'`;
 }
 
 // A vega-lite chart displayed in the footer of a <StatisticBox />.
@@ -51,16 +34,13 @@ export default function FooterChart(props: Props) {
         {({ width }) => {
           if (width > 0) {
             return (
-              <VegaLite
+              <BaseChart
                 width={width}
                 height={height}
-                renderer="canvas"
-                theme="dark"
-                tooltip={{theme: 'dark'}}
-                actions={false}
                 data={{default_data: props.data}}
                 config={props.config || defaultConfig}
-                spec={spec as VisualizationSpec} />
+                spec={spec as VisualizationSpec}
+              />
             );
           }
           return null;
