+++ src/parser/shared/modules/features/RaidHealthTab/TabComponent/RaidHealthChart.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { VegaLite } from 'react-vega';
 import { AutoSizer } from 'react-virtualized';
-import { formatTime, defaultConfig } from 'interface/others/FooterChart';
+import BaseChart, { formatTime } from 'interface/others/BaseChart';
 
 const DEATH_COLOR = 'rgba(255, 0, 0, 0.8)';
 
@@ -124,15 +123,11 @@ class RaidHealthChart extends React.Component {
     return (
       <AutoSizer disableHeight>
         {({ width }) => (
-          <VegaLite
+          <BaseChart
             height={400}
             width={width}
-            config={defaultConfig}
             spec={spec}
             data={data}
-            actions={false}
-            theme="dark"
-            tooltip={{theme: 'dark'}}
           />
         )}
       </AutoSizer>
